store/display data from quotes json
