+++ src/objects/quotes.json
@@ -1,0 +1,23 @@
+{
+    "quote1": "The only way to do great work is to love what you do. - Steve Jobs",
+    "quote2": "In the middle of difficulty lies opportunity. - Albert Einstein",
+    "quote3": "Be yourself; everyone else is already taken. - Oscar Wilde",
+    "quote4": "You miss 100% of the shots you don’t take. - Wayne Gretzky",
+    "quote5": "Believe you can and you're halfway there. - Theodore Roosevelt",
+    "quote6": "The only thing we have to fear is fear itself. - Franklin D. Roosevelt",
+    "quote7": "Life is what happens when you're busy making other plans. - John Lennon",
+    "quote8": "The only thing we have to fear is fear itself. - Franklin D. Roosevelt",
+    "quote9":  "Don't watch the clock; do what it does. Keep going. - Sam Levenson",
+    "quote10": "Change your thoughts and you change your world. - Norman Vincent Peale",
+    "quote11": "Happiness depends upon ourselves. - Aristotle",
+    "quote12": "The best way to predict the future is to create it. - Peter Drucker",
+    "quote13": "It does not matter how slowly you go as long as you do not stop. - Confucius",
+    "quote14": "The only way to do great work is to love what you do. - Steve Jobs",
+    "quote15": "We may encounter many defeats but we must not be defeated. - Maya Angelou",
+    "quote16": "Strive not to be a success, but rather to be of value. - Albert Einstein",
+    "quote17": "You must be the change you wish to see in the world. - Mahatma Gandhi",
+    "quote18": "I have not failed. I've just found 10,000 ways that won't work. - Thomas Edison",
+    "quote19": "Strive not to be a success, but rather to be of value. - Albert Einstein",
+    "quote20": "Don't watch the clock; do what it does. Keep going. - Sam Levenson"
+}
+
