+++ src/tabs/newtab.jsx
@@ -6,7 +6,9 @@ import RandomWallpaper from "../components/wallpaper/RandomWallpaper.jsx";
 import Geolocation from "../components/geolocation/Geolocation.jsx";
 import Message from '../components/searchBar/message.jsx';
 import TodoApp from '../components/todoList/todoList.jsx'
-import Quote from "../components/quotes/Quote.jsx";
+import RandomQuote from "../components/quotes/Quote.jsx";
+
+
 
 function Tab() {
   return (
@@ -16,8 +18,8 @@ function Tab() {
       <Message />
       <Weather />
       <TodoApp />
-      <Quote/>
       <Geolocation />
+      <RandomQuote/>   
     </div>
     </>
   );
